`Pages::ShowView::BreadcrumbsComponent` -> `Breadcrumbs::TopicBreadcrumbsComponent`
