+++ app/views/pages/show_view.html.erb
@@ -2,7 +2,7 @@
   <%= layout.with_header do %>
     <%= render Headers::GlobalComponent.new(current_user:, current_page_name:) do |header| %>
       <%= header.with_breadcrumbs do %>
-        <%= render Pages::ShowView::BreadcrumbsComponent.new(topic:) %>
+        <%= render Breadcrumbs::TopicBreadcrumbsComponent.new(topic:) %>
       <% end %>
     <% end %>
   <% end %>
