--- app/views/pages/show_view/breadcrumbs_component.rb
@@ -1,19 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-module Pages
-  class ShowView
-    class BreadcrumbsComponent < ApplicationComponent
-      sig { params(topic: Topic).void }
-      def initialize(topic:)
-        @topic = topic
-      end
-
-      sig { returns(Topic) }
-      attr_reader :topic
-      private :topic
-
-      delegate :space, to: :topic
-    end
-  end
-end
